requirements.txt
Django==3.1.4
django-environ==0.4.5
psycopg2==2.8.6
gunicorn==20.0.4
django-compressor==2.4
django-htmlmin==0.11.0
django-debug-toolbar==3.2
django-admin-autocomplete-filter==0.6.1
django-extensions==3.1.0
django-crispy-forms==1.10.0
django-uuslug==1.2.0
Pillow==8.0.1
tensorflow==2.4.0
numpy==1.19.4
